delete-tratado

+++ src/hooks/contexts/CandidatosContext.tsx
@@ -15,9 +15,9 @@ type CandidatosContextProps = {
 const CandidatosContext = createContext<CandidatosContextType>(
     {
         candidatos: [],
-        postCandidato: async () => {},
-        putCandidato: async () => {},
-        deleteCandidato: async () => {},
+        postCandidato: async () => { },
+        putCandidato: async () => { },
+        deleteCandidato: async () => { },
     }
 );
 
@@ -60,18 +60,21 @@ export const CandidatosContextProvider = ({ children }: CandidatosContextProps) 
         }
     };
 
-      // DELETE
-      const deleteCandidato = async (id: number) => {
-        try {
-            await axios.delete(`http://localhost:3000/candidatos/${id}`);
-            fetchCandidatos(); // Atualiza a lista após a exclusão
-        } catch (error) {
-            console.log("Erro ao deletar candidato:", error);
+    // DELETE
+    const deleteCandidato = async (id: number) => {
+        const confirmDelete = confirm("Tem certeza que deseja ELIMINAR o candidato?");
+        if (confirmDelete) {
+            try {
+                await axios.delete(`http://localhost:3000/candidatos/${id}`);
+                fetchCandidatos(); // Atualiza a lista após a exclusão
+            } catch (error) {
+                console.log("Erro ao deletar candidato:", error);
+            }
         }
     };
 
     return (
-        <CandidatosContext.Provider value={{ candidatos, postCandidato, putCandidato, deleteCandidato}}>
+        <CandidatosContext.Provider value={{ candidatos, postCandidato, putCandidato, deleteCandidato }}>
             {children}
         </CandidatosContext.Provider>
     );
